+++ lib/metrics_analyzer.dart
@@ -1,3 +1,2 @@
 export 'package:dart_code_metrics/src/config_builder/models/config.dart';
-export 'package:dart_code_metrics/src/analyzers/lint_analyzer/constants.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/lint_analyzer.dart';