+++ test/analyzers/lint_analyzer/anti_patterns/anti_patterns_list/long_method/long_method_test.dart
@@ -1,7 +1,7 @@
 @TestOn('vm')
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:dart_code_metrics/src/analyzers/lint_analyzer/anti_patterns/anti_patterns_list/long_method.dart';
-import 'package:dart_code_metrics/src/analyzers/lint_analyzer/constants.dart';
+import 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/metrics_list/source_lines_of_code/source_lines_of_code_metric.dart';
 import 'package:dart_code_metrics/src/analyzers/lint_analyzer/scope_visitor.dart';
 import 'package:dart_code_metrics/src/analyzers/models/severity.dart';
 import 'package:test/test.dart';
@@ -31,7 +31,7 @@ void main() {
 
         return true;
       }),
-      {linesOfExecutableCodeKey: 25},
+      {SourceLinesOfCodeMetric.metricId: 25},
     );
 
     AntiPatternTestHelper.verifyInitialization(
@@ -47,10 +47,10 @@ void main() {
       startColumns: [1],
       endOffsets: [1309],
       messages: [
-        'Long function. This function contains 29 lines with executable code.',
+        'Long function. This function contains 29 lines with code.',
       ],
       verboseMessage: [
-        "Based on configuration of this package, we don't recommend write a function longer than 25 lines with executable code.",
+        "Based on configuration of this package, we don't recommend write a function longer than 25 lines with code.",
       ],
     );
   });