+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Remove deprecated `Lines of Executable Code` metric, use `Source lines of Code` instead.
+
 ## 3.2.1
 
 * Remove unnecessary scan by `Lines of Executable Code`